\name{dropbox.file.info}
\alias{dropbox.file.info}
\title{Return file attributes for a specified file supplied in the path argument.}
\usage{
  dropbox.file.info(cred, path_to_file)
}
\arguments{
  \item{cred}{An object of class ROAuth with Dropobox
  specific credentials.}

  \item{path_to_file}{<what param does>}
}
\value{
  list
}
\description{
  Return file attributes for a specified file supplied in
  the path argument.
}
\examples{
\dontrun{

}
}
\seealso{
  is.dropbox.file is.dropbox.dir
}

