\name{is.valid.dropbox.operation}
\alias{is.valid.dropbox.operation}
\title{Function to handle errors if a returned object is not the excepted JSON object}
\usage{
  is.valid.dropbox.operation(dropbox_call)
}
\arguments{
  \item{dropbox_call}{A function call to a Dropbox method
  via OAuth$handshake()}
}
\value{
  logical
}
\description{
  Function to handle errors if a returned object is not the
  excepted JSON object
}
\examples{
\dontrun{
example forthcoming
}
}

