\name{is.valid.revision}
\alias{is.valid.revision}
\title{Checks whether supplied revision number is valid on Dropobx}
\usage{
  is.valid.revision(cred, path = NULL, revision = NULL)
}
\arguments{
  \item{cred}{An object of class ROAuth with Dropobox
  specific credentials.}

  \item{path}{path to file or folder. Full path if
  file/folder is not in Dropbox root.}

  \item{revision}{revision number}
}
\value{
  logical
}
\description{
  Checks whether supplied revision number is valid on
  Dropobx
}
\examples{
\dontrun{
Not yet coded.
}
}

